Move State into new module
